Reorganization of campaigns code
