+++ client/mainmenu/CPrologEpilogVideo.cpp
@@ -18,8 +18,6 @@
 #include "../widgets/TextControls.h"
 #include "../render/Canvas.h"
 
-#include "../../lib/mapping/CCampaignHandler.h"
-
 
 CPrologEpilogVideo::CPrologEpilogVideo(CampaignScenarioPrologEpilog _spe, std::function<void()> callback)
 	: CWindowObject(BORDERED), spe(_spe), positionCounter(0), voiceSoundHandle(-1), exitCb(callback)
