+++ lib/gameState/CGameState.cpp
@@ -26,13 +26,13 @@
 #include "../VCMIDirs.h"
 #include "../VCMI_Lib.h"
 #include "../battle/BattleInfo.h"
+#include "../campaign/CampaignState.h"
 #include "../filesystem/ResourceID.h"
 #include "../mapObjectConstructors/AObjectTypeHandler.h"
 #include "../mapObjectConstructors/CObjectClassesHandler.h"
 #include "../mapObjectConstructors/DwellingInstanceConstructor.h"
 #include "../mapObjects/CGHeroInstance.h"
 #include "../mapObjects/CGTownInstance.h"
-#include "../mapping/CCampaignHandler.h"
 #include "../mapping/CMap.h"
 #include "../mapping/CMapEditManager.h"
 #include "../mapping/CMapService.h"
