+++ lib/registerTypes/RegisterTypes.cpp
@@ -16,7 +16,6 @@
 #include "../mapObjects/CObjectHandler.h"
 #include "../CCreatureHandler.h"
 #include "../spells/CSpellHandler.h"
-#include "../mapping/CCampaignHandler.h"
 
 #include "../serializer/BinaryDeserializer.h"
 #include "../serializer/BinarySerializer.h"
