+++ server/CGameHandler.cpp
@@ -12,7 +12,6 @@
 #include "../lib/filesystem/Filesystem.h"
 #include "../lib/filesystem/FileInfo.h"
 #include "../lib/int3.h"
-#include "../lib/mapping/CCampaignHandler.h"
 #include "../lib/ArtifactUtils.h"
 #include "../lib/StartInfo.h"
 #include "../lib/CModHandler.h"
